Test debug all examples

+++ src/node/Mesh.cpp
@@ -116,8 +116,8 @@ void Mesh::setVertexBuffer() {
     conversorVBO(vertexIndex, vertexList, normalIndex, normalList, textureIndex, textureList, vertexDataIn);
     indexVBO_slow(vertexDataIn, vertexData, indexIBO);
 
-    SDL_LogDebug(SDL_LOG_CATEGORY_RENDER, "VBO Nome: %s Indice: %d Data: %d", getName().c_str(), indexIBO.size(),
-                 vertexData.size());
+    SDL_LogDebug(SDL_LOG_CATEGORY_RENDER, "VBO Nome: %s Indice: %d Data: %d", getName().c_str(), (int)indexIBO.size(),
+                 (int)vertexData.size());
 
     unsigned int sizeBufferVertex = vertexData.size() * sizeof(VertexData);
     unsigned int sizeBufferIndex = indexIBO.size() * sizeof(unsigned int);
